- some changes for adding appstates on the fly

+++ MARS_NB/MARS/MARSCore/src/mars/actuators/BallastTank.java
@@ -230,7 +230,7 @@ public class BallastTank extends Actuator implements Keys,ChartValue{
             }
             //System.out.println("desired: " + getDesiredVolume() + " current" + currentVolume);
             float buoyancy_force = getPhysical_environment().getFluid_density() * getPhysical_environment().getGravitational_acceleration() * getCurrentVolume();
-            physics_control.applyImpulse(Vector3f.UNIT_Y.negate().mult(buoyancy_force/simauv_settings.getPhysicsFramerate()), this.getMassCenterGeom().getWorldTranslation().subtract(BallastStart.getWorldTranslation()));
+            physics_control.applyImpulse(Vector3f.UNIT_Y.negate().mult(buoyancy_force/((float)simauv_settings.getPhysicsFramerate())), this.getMassCenterGeom().getWorldTranslation().subtract(BallastStart.getWorldTranslation()));
         //}
     }
